requirements.txt
python-telegram-bot==20.3
openai==1.12.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1